БАГИ ФИКС АВТОРИЗ
