+++ src/layout/Auth.css
@@ -1,0 +1,39 @@
+
+    body, html {
+      margin: 0;
+      padding: 0;
+      height: 100%;
+      font-family: Arial, sans-serif;
+    }
+
+    .container {
+      display: flex;
+      height: 100vh;
+    }
+
+    .left-column {
+      flex: 1;
+      background-color: #f0f0f0;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      overflow: hidden;
+    }
+
+    .left-column img {
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+    }
+
+    .right-column {
+      flex: 2;
+      background-color: #ffffff;
+      padding: 40px;
+      overflow-y: auto;
+    }
+
+    .app-content {
+      max-width: 600px;
+      margin: auto;
+    }
