--- src/pages/Error.tsx
@@ -1,3 +1,0 @@
-export default function Error () {
-    return <h1>Ошибка</h1>
-}
