+++ src/pages/TodoPage.tsx
@@ -1,13 +1,13 @@
 import { useQuery } from "@tanstack/react-query";
-import { getTasks } from "../Api/apiTasksTodo.js";
+import { getTasks } from "../api/todos.js";
 import TodoList from "../Components/TodoList.js";
 import AddTask from "../Components/AddTask.js";
 import TabsList from "../Components/TabsList.js";
-import { Todo, TodoInfo } from "../types/type.js";
+import { Todo, TodoInfo } from "../types/todos.js";
 import { setStatus } from "../store/Slice.js";
 import { useDispatch, useSelector } from "react-redux";
 
-export default function TodoPage() {
+export default function TodoPage(): JSX.Element {
   const dispatch = useDispatch();
   const currentStatusType = useSelector(
     (state: any) => state.status.currentStatusType
